+++ src/EventSystem/Response/EventListener/EtagResponseEventListener.php
@@ -4,6 +4,7 @@ namespace App\EventSystem\Response\EventListener;
 
 use App\Response\CollectionResponse;
 use App\Response\ElementResponse;
+use App\Response\NotModifiedResponse;
 use App\Service\EtagService;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 
@@ -17,7 +18,11 @@ class EtagResponseEventListener
     public function onKernelResponse(ResponseEvent $event): void
     {
         $response = $event->getResponse();
-        if (!($response instanceof CollectionResponse || $response instanceof ElementResponse)) {
+        if (!(
+            $response instanceof CollectionResponse
+            || $response instanceof ElementResponse
+            || $response instanceof NotModifiedResponse
+        )) {
             return;
         }
         $etag = $this->etagService->getCurrentRequestEtag();